refactor: shared folders plural
