+++ src/shared/configs/configuration.ts
