+++ src/shared/configs/module-options.ts
