+++ src/shared/middlewares/request-id/request-id.middleware.ts
@@ -1,4 +1,4 @@
-import { REQUEST_ID_TOKEN_HEADER } from './../../constant';
+import { REQUEST_ID_TOKEN_HEADER } from '../../constants';
 import { Request, Response } from 'express';
 import { v4 as uuidv4 } from 'uuid'
