+++ src/user/controllers/user.controller.ts
@@ -10,7 +10,7 @@ import {
 import { Request } from 'express';
 
 import { UserService } from '../services/user.service';
-import { GetMeOutput } from '../dto/me.dto';
+import { GetMeOutput } from '../dtos/me.dto';
 import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
 
 @Controller('users')
